--- components/SEO.tsx
@@ -1,64 +1,0 @@
-import Head from "next/head"
-
-/**
- * Provide SEO related meta tags to a page.
- *
- * @param {Props} props The props object.
- * @param {string} props.title Used for the page title, og:title, twitter:title, etc.
- * @param {string} props.description Used for the meta description, og:description, twitter:description, etc.
- * @param {string} props.imageUrl Used for the og:image and twitter:image.
- * @param {string} props.url Used for the og:url and twitter:url.
- *
- * @returns {React.ReactElement} The SEO component
- */
-
-interface SEOProps {
-  title?: string
-  description?: string
-  imageUrl?: string
-  url?: string
-}
-
-export default function SEO({ title, description, imageUrl, url }: SEOProps) {
-  if (!title && !description && !imageUrl && !url) {
-    return null
-  }
-
-  return (
-    <Head>
-      <meta property="og:type" content="website" />
-      <meta property="twitter:card" content="summary_large_image" />
-
-      {title && (
-        <>
-          <title>{title}</title>
-          <meta name="title" content={title} />
-          <meta property="og:title" content={title} />
-          <meta property="twitter:title" content={title} />
-        </>
-      )}
-
-      {description && (
-        <>
-          <meta name="description" content={description} />
-          <meta property="og:description" content={description} />
-          <meta property="twitter:description" content={description} />
-        </>
-      )}
-
-      {imageUrl && (
-        <>
-          <meta property="og:image" content={imageUrl} />
-          <meta property="twitter:image" content={imageUrl} />
-        </>
-      )}
-
-      {url && (
-        <>
-          <meta property="og:url" content={url} />
-          <meta property="twitter:url" content={url} />
-        </>
-      )}
-    </Head>
-  )
-}
