--- pages/[...wordpressNode].tsx
@@ -1,16 +1,0 @@
-import { getWordPressProps, WordPressTemplate } from '@faustwp/core';
-
-export default function Page(props) {
-  return <WordPressTemplate {...props} />;
-}
-
-export function getStaticProps(ctx) {
-  return getWordPressProps({ ctx });
-}
-
-export async function getStaticPaths() {
-  return {
-    paths: [],
-    fallback: 'blocking',
-  };
-}
