--- pages/_app.tsx
@@ -1,15 +1,0 @@
-import '../faust.config';
-import React from 'react';
-import { useRouter } from 'next/router';
-import { FaustProvider } from '@faustwp/core';
-import '../styles/global.scss';
-
-export default function MyApp({ Component, pageProps }) {
-  const router = useRouter();
-
-  return (
-    <FaustProvider pageProps={pageProps}>
-      <Component {...pageProps} key={router.asPath} />
-    </FaustProvider>
-  );
-}
