--- pages/index.tsx
@@ -1,9 +1,0 @@
-import { getWordPressProps, WordPressTemplate } from '@faustwp/core';
-
-export default function Page(props) {
-  return <WordPressTemplate {...props} />;
-}
-
-export function getStaticProps(ctx) {
-  return getWordPressProps({ ctx });
-}
