--- pages/preview.tsx
@@ -1,5 +1,0 @@
-import { WordPressTemplate } from '@faustwp/core';
-
-export default function Preview(props) {
-  return <WordPressTemplate {...props} />;
-}
