+++ src/components/Header/Logo.tsx
@@ -1,6 +1,6 @@
-import Image from "next/image"
-import logo from "@core/public/images/logo.png"
-import style from "@core/styles/components/Header/Logo.module.scss"
+import Image from "next/image";
+import logo from "@core/public/images/logo.png";
+import style from "@core/src/styles/components/Header/Logo.module.scss";
 
 export default function HeaderLogo() {
 	return (
@@ -8,5 +8,5 @@ export default function HeaderLogo() {
 			<Image src={logo.src} width={40} height={40} alt="Bartr Logo" />
 			<span>Bartr</span>
 		</div>
-	)
+	);
 }
