+++ src/components/Header/Main.tsx
@@ -1,8 +1,8 @@
-import { MenuItem } from "@core/constants/interfaces"
-import Link from "next/link"
-import style from "@core/styles/components/Header/Main.module.scss"
-import HeaderCTA from "@core/components/Header/CTA"
-import HeaderLogo from "@core/components/Header/Logo"
+import { MenuItem } from "@core/src/constants/interfaces";
+import Link from "next/link";
+import style from "@core/src/styles/components/Header/Main.module.scss";
+import HeaderCTA from "@core/src/components/Header/CTA";
+import HeaderLogo from "@core/src/components/Header/Logo";
 
 export default function Header({ menuItems }: { menuItems: MenuItem[] }) {
 	return (
@@ -15,11 +15,11 @@ export default function Header({ menuItems }: { menuItems: MenuItem[] }) {
 						<Link key={key} href={item.url} target={item.target}>
 							<a className={item.cssClasses.join(" ")}>{item.label}</a>
 						</Link>
-					)
+					);
 				})}
 			</div>
 
 			<HeaderCTA />
 		</header>
-	)
+	);
 }
