+++ src/constants/interfaces.ts
