+++ src/pages/api/faust/[[...route]].js
