+++ src/pages/preview.tsx
@@ -1,0 +1,5 @@
+import { WordPressTemplate } from "@faustwp/core";
+
+export default function Preview(props) {
+	return <WordPressTemplate {...props} />;
+}
