--- styles/components/Header/Logo.module.scss
@@ -1,11 +1,0 @@
-@import "@core/styles/_mixins.scss";
-@import "@core/styles/_brand-variables.scss";
-
-.logo {
-	@include flex(row);
-	gap: 0;
-
-	span {
-		color: $secondary;
-	}
-}
